pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-agent"
version = "0.1.0"
description = "AI-powered email management agent with smart triage, categorization, and reply drafting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [{name = "bludragon-ai"}]
dependencies = [
    "langchain>=0.3,<0.4",
    "langchain-anthropic>=0.3,<0.4",
    "langchain-openai>=0.3,<0.4",
    "streamlit>=1.40,<2",
    "pydantic>=2.0,<3",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
